requirements.txt
fastapi==0.104.1
faster-whisper==0.10.1
pycryptodome==3.19.0
python-multipart==0.0.6
uvicorn[standard]==0.24.0